requirement.txt
langchain
python-dotenv
ipykernel
langchain_groq
langchain_core
fastapi
uvicorn
langserve
pandas
langchain-groq
str
streamlit
pypdf
